refactor: deleted utils/utils.go and move BitbucketPagination from api_client.go to api_common.go
